erpAgroflori/requirements.txt
django
django-environ
django-crispy-forms
Pillow
django-phonenumber-field[phonenumbers]
psycopg2
crispy_bootstrap4
django-allauth
plotly